+++ src/main/java/com/flather/weatherstation/dto/weather/WeatherRecordCreatedDto.java
@@ -17,7 +17,7 @@ public class WeatherRecordCreatedDto {
 
   private Double humidity;
 
-  private Long surfaceWetness;
+  private Double surfaceWetness;
 
   private double WIFI_RSSI;
 }
